requirements.txt
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
html2text==2020.1.16
idna==3.4
Mastodon.py==1.8.0
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.28.2
six==1.16.0
urllib3==1.26.15
